osa5 5.5-5.11

+++ osa4-5/blogilista/backend/controllers/blogs.js
@@ -48,8 +48,8 @@ blogsRouter.get("/:id", async (request, response) => {
 
 blogsRouter.delete("/:id", userAthorization, async (request, response) => {
   const blog = await Blog.findById(request.params.id);
-
-  if (blog.user.toString() === request.user.id.toString()) {
+  console.log(request.user);
+  if (blog.user && blog.user.toString() === request.user.id.toString()) {
     await Blog.findByIdAndDelete(blog._id);
     response.status(204).end();
   }
@@ -63,17 +63,54 @@ blogsRouter.delete("/:id", userAthorization, async (request, response) => {
 blogsRouter.put("/:id", userAthorization, async (request, response) => {
   const blog = await Blog.findById(request.params.id);
 
+  if (!blog) {
+    return response.status(404).end();
+  }
+
   if (blog.user?.toString() !== request.user.id.toString()) {
-    response
+    return response
       .status(403)
       .json({ error: "only the creator can edit this blog" })
       .end();
   }
 
-  blog.likes = request.body.likes;
+  const { title, author, url, likes } = request.body;
+
+  blog.title = title;
+  blog.author = author;
+  blog.url = url;
+  blog.likes = likes ?? blog.likes ?? 0;
 
   const updatedBlog = await blog.save();
-  response.status(200).json(updatedBlog);
+  return response.status(200).json(updatedBlog);
+});
+
+blogsRouter.patch("/like/:id", userAthorization, async (request, response) => {
+  const { type } = request.body ?? {};
+
+  if (!["like", "dislike"].includes(type)) {
+    return response.status(400).json({
+      error: "type must be like or dislike",
+    });
+  }
+  const amount = type === "like" ? 1 : -1;
+
+  const updatedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    { $inc: { likes: amount } },
+    {
+      new: true,
+      runValidators: true,
+    },
+  ).populate("user", {
+    name: 1,
+    username: 1,
+  });
+
+  if (!updatedBlog) {
+    return response.status(404).end();
+  }
+  return response.status(200).json(updatedBlog);
 });
 
 module.exports = blogsRouter;
